fix: missing method error
