+++ main.js
@@ -15,6 +15,7 @@ import get_env from './get_env'
 
 import ServerRouter from './server'
 import schema from './template/schema'
+const ohyaySchema = require('./server/streaming/ohyay/src/schema/schema')
 import TemplateRouter from './template'
 const app = express()
 
@@ -199,6 +200,29 @@ const apolloserver = new ApolloServer({
 })
 
 apolloserver.applyMiddleware({ app })
+
+// ohyay remote schema integration
+const ohyayApolloserver = new ApolloServer({
+   schema: ohyaySchema,
+   playground: {
+      endpoint: `${get_env('ENDPOINT')}/ohyay/graphql`
+   },
+   introspection: true,
+   validationRules: [depthLimit(11)],
+   formatError: err => {
+      console.log(err)
+      if (err.message.includes('ENOENT'))
+         return isProd ? new Error('No such folder or file exists!') : err
+      return isProd ? new Error(err) : err
+   },
+   debug: true,
+   context: ({ req }) => {
+      const ohyay_api_key = req.header('ohyay_api_key')
+      return { ohyay_api_key }
+   }
+})
+
+ohyayApolloserver.applyMiddleware({ app })
 
 app.listen(PORT, () => {
    console.log(`Server started on ${PORT}`)
